--- calpool-fe/src/components/Button/Button.css
@@ -1,22 +1,0 @@
-/* Bttn.css */
-.button-custom {
-  /* Default values */
-
-  padding: 10px 30px;
-  text-transform: none;
-  border: none;
-  border-radius:.25rem;
-  color: black;
-  background-color: gray;
-  transition: filter 100ms ease-in-out;
-  font-size: 1.1rem;
-  font-weight: 600;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.button-custom:hover {
-  cursor: pointer;
-  filter: brightness(85%);
-}
